+++ apps/api/src/kick_logs/domain/entities/__init__.py
@@ -1,1 +1,7 @@
-"""Domain entities."""
+from kick_logs.domain.entities.channel import Channel
+from kick_logs.domain.entities.chat_message import ChatMessage
+from kick_logs.domain.entities.emote import Emote
+from kick_logs.domain.entities.sender import Sender
+from kick_logs.domain.entities.user import User
+
+__all__ = ["Channel", "ChatMessage", "Emote", "Sender", "User"]